+++ laser.cpp
@@ -1,7 +1,8 @@
 #include "laser.h"
 
 namespace game {
-	Laser::Laser(const std::string name, const Resource *geometry, const Resource *material, const Resource *tex) : SceneNode(name, geometry, material, tex) {}
+	Laser::Laser(const std::string name, const Resource *geometry, 
+		const Resource *material, const Resource *tex) : SceneNode(name, geometry, material, tex) {}
 
 	Laser::~Laser() {}
 
@@ -13,7 +14,7 @@ namespace game {
 		direction = dir;
 	}
 
-	void Laser::Update(float delta_time) {
+	void Laser::Update(double delta_time) {
 		glm::quat rotation = glm::angleAxis(glm::pi<float>() / 180.0f, glm::vec3(0.0, 0.0, 1.0));
 		this->SetOrientation(this->GetOrientation() * rotation); //laser cube spins as it flies away to give more of a "lasery" impression
